updated basicSchedule to version 0.21
Continued work on the buildSchedule function. Changed any reference to slot size to refer to number of slots rather than length in minutes to be uniform and hopefully simpler.  So far ties for longest on break are simply being handled as the first person to occupy the longest spot will get the next position. We will increase complexity here later by looking at what positions they've already worked and try to balance it.

Need to continue work on this function, specifically the part where we assign an employee to a position.  Need to figure out all the actions that need to be taken and see if there's a way to shorten to code for this.